Code cleanup of MissionCheats
